Fix `removeListeners()` removing unrelated events
This fixes a long-running issue that made `removeListeners()` remove all
events under a given namespace if there were other events that whose
names ended with the same string. For instance, if a `foobar.scroll`
event was registered and `removeListeners()` was called to remove
`bar.scroll` events, `foobar.scroll` would also be removed even it not
being supposed to.