Make the Reveal plugin smarter
This adds loads of new methods to make the plugin easier to read and
maintain. It also removes the need to use the container
`c-reveal__content` and the class `c-reveal--hidden`. From this point on
the plugin will always ask which dimensions it should animate instead of
inferring that only once at construction time. This makes it possible to
run the plugin even if the container was initially hidden with
`display:none`.