Change window scroll to page offset
This fixes bugs happening in IE.
pageYOffset and pageXOffset are read-only window properties that return
the number of pixels the document has been scrolled vertically.
And scrollY and scrollX are the exactly same thing. The main difference
is that there is better cross-browser compatibility for
page[axis]Offset than scroll[axis].